policy-navigator/app/requirements.txt
google-cloud-aiplatform[adk, agent_engines]
python-dotenv
google-genai
google-adk
asgiref==3.8.1
cloudpickle==3.1.1
beautifulsoup4
Markdown==3.8
pydantic==2.12.2